+++ ChangeLog.md
@@ -3,6 +3,7 @@
 * Fix exception when disposing DicomDirectory
 * Fix menu bug in DICOM Dump utility when changing syntax to JPEG Lossless
 * Ability to browse multiframe images in DICOM Dump utility
+* Add classes for basic structured report creation
 
 #### v1.0.33 (4/28/2013)
 * Add logging abstraction layer to remove runtime dependency on NLog